apply spotless

+++ src/main/java/com/sinthoras/visualprospecting/Utils.java
@@ -1,6 +1,11 @@
 package com.sinthoras.visualprospecting;
 
-import java.io.*;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.Reader;
+import java.io.Writer;
 import java.nio.ByteBuffer;
 import java.nio.channels.FileChannel;
 import java.nio.file.Files;
@@ -205,8 +210,7 @@ public class Utils {
 
     public static void writeToFile(File file, ByteBuffer byteBuffer) {
         try {
-            if (file.exists())
-                file.delete();
+            if (file.exists()) file.delete();
 
             final FileOutputStream outputStream = new FileOutputStream(file, false);
             final FileChannel outputChannel = outputStream.getChannel();
